feat:商品服务增加必要参数空指针判断

+++ module-management-service/module-service-product/src/main/java/com/zn/gmall/product/controller/admin/BaseCategoryTrademarkController.java
@@ -6,6 +6,7 @@ import com.zn.gmall.model.product.vo.CategoryTrademarkVo;
 import com.zn.gmall.product.service.api.BaseCategoryTrademarkService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,6 +18,7 @@ import java.util.List;
 @Api(tags = "分类品牌列表接口")
 @RestController
 @RequestMapping("admin/product/baseCategoryTrademark")
+@Slf4j
 public class BaseCategoryTrademarkController {
 
     @Autowired
@@ -30,6 +32,10 @@ public class BaseCategoryTrademarkController {
     @ApiOperation("保存分类与品牌关联")
     @PostMapping("save")
     public Result<Void> save(@RequestBody CategoryTrademarkVo categoryTrademarkVo) {
+        log.info("保存分类与品牌关联:{}", categoryTrademarkVo);
+        if (categoryTrademarkVo == null) {
+            return Result.<Void>fail().message("参数不能为空");
+        }
         //  保存方法
         baseCategoryTrademarkService.save(categoryTrademarkVo);
         return Result.ok();
@@ -44,6 +50,10 @@ public class BaseCategoryTrademarkController {
     @ApiOperation("删除关联")
     @DeleteMapping("remove/{category3Id}/{trademarkId}")
     public Result<Void> remove(@PathVariable Long category3Id, @PathVariable Long trademarkId) {
+        log.info("删除关联:category3Id:{},trademarkId:{}", category3Id, trademarkId);
+        if (category3Id == null || trademarkId == null) {
+            return Result.<Void>fail().message("品牌id/三级分类id不能为空");
+        }
         //  调用服务层方法
         baseCategoryTrademarkService.remove(category3Id, trademarkId);
         return Result.ok();
@@ -58,6 +68,10 @@ public class BaseCategoryTrademarkController {
     @ApiOperation("根据三级分类获取品牌")
     @GetMapping("findTrademarkList/{category3Id}")
     public Result<List<BaseTrademark>> findTrademarkList(@PathVariable Long category3Id) {
+        log.info("根据三级分类获取品牌:category3Id:{}", category3Id);
+        if (category3Id == null) {
+            return Result.<List<BaseTrademark>>fail().message("三级分类id不能为空");
+        }
         //  select * from base_trademark
         List<BaseTrademark> list = baseCategoryTrademarkService.findTrademarkList(category3Id);
         //  返回
@@ -73,6 +87,10 @@ public class BaseCategoryTrademarkController {
     @ApiOperation("获取当前未被三级分类关联的所有品牌")
     @GetMapping("findCurrentTrademarkList/{category3Id}")
     public Result<List<BaseTrademark>> findCurrentTrademarkList(@PathVariable Long category3Id) {
+        log.info("获取当前未被三级分类关联的所有品牌:category3Id:{}", category3Id);
+        if (category3Id == null) {
+            return Result.<List<BaseTrademark>>fail().message("三级分类id不能为空");
+        }
         List<BaseTrademark> list = baseCategoryTrademarkService.findCurrentTrademarkList(category3Id);
         //  返回
         return Result.ok(list);
